Remove coupling with the controller and but keep running the block only once
Before the change on #36, the block was only executed once in case the
lock was acquired but the block returned `nil`. With this change we go
back to that behavior but we don't need to check for the controller or
the response state like it was before #36.

This is an alternative to #38 and #39.